Replace revision with KB number
The revision number used to identify which updates had been installed
assumed every KB would increment the revision, however, this is not
true. KBs that do not update the kernel may not change the revision
number.

KBs that don't change the revision number are still important to track,
since it is only user-mode components that are logging interesting Event
Log events in the first place. Luckily, since Windows 10 updates are
cumulative, tracking the newest KB installed before each collected event
uniquely identifies all the fixes that are in effect at the time the
event was logged.

Installed KBs are located via the WMI Win32_QuickFixEngineering provider
and the Setup log. The MI .Net client is used to query WMI, which
requires WinRM to be running on the machine and the Windows 10 SDK to be
installed.

Also, build number changes are tracked via EventLog event 6009.

Switching from revision to KB is a breaking change to the file format,
so the assembly version is incremented. The analysis tool was updated
accordingly, but not compiled or tested because it is not compatible
with VS 2013.